Implement login and register
When not logged in you will be redirected to the login page, and from
there you can go to the register page to create a new user.

As of now, it will set the state in App.tsx, and pass the information
about the user down to the children by using React.Context.

Tomorrow, I will rework how the tokens are stored - probably using
cookies, as this is better for the user, and from a security standpoint.