Add ability to log out and recover session on reenter
When you open the app now after you've been logged in previously, it
will try to refetch the token from localStorage, and request the user.
Thus you don't have to log onto the app every time you open it.

In the future, we have to check if the token has expired, and use a
refresh token to get a new one, and if that again fails, we have to log
out the user.