Implement balance API
This is not an optimal implementation, but it works. In future
versions we should try to leverage the database and cache more.

The BalanceByDateRangeView doesn't meet the spec, as it returnes a
list instead of a pagination object. This is because the standard
Django Rest Framework pagination only works with querysets. To use
it with generated objects (which is what balances are) we have to
write a custom pagination handler.

The Balance class should probably not be in balance/serializers.py,
but I didn't know where else to put it.

Fixes #10