All reducer actions now use [...]Actions.do[...] pattern
Fixes #91.

The reason why I removed a lot of comments are because they didn't add
anything to the code. They simply took up space. If someone wants to
look at our reducer code, they should already understand what an action
is, and how a reducer works. It is not our task to teach them that, and
thus we shouldn't have comments that explain that. A lot of them felt
like they were simply there for the sake of having a comment there,
which I think is strictly unnecessary, and they only make it harder to
read the file - contrary to its purpose.