Fix auth test and begin setting up standalone test helper
Spent A LOT of time figuring out issues with PUT and DELETE `/user/`.
Turns out that ?email=... doesn't do anything for DELETE, which led to a
horrible situation where no error codes made any sense.

Have I torn my hair out for three hours straight? Yes.

Did I stop when pomodoro told me to? No.

Did I fix it? You betcha.

The changes should hopefully soon be reflected in swagger and the
backend, but for now, just keep this in mind.

There is also an issue with PUT `/user/` which we have not figured out
yet, but it should be resolved shortly.